Still working on creating an order

+++ src/main/java/ch/roester/shipping_method/ShippingMethodRequestDTO.java
@@ -20,6 +20,6 @@ public class ShippingMethodRequestDTO {
 
     private String description;
 
-    private Integer weightInGramsLimit;
+    private BigDecimal weightInGramsLimit;
 
 }
